Adapted preprocessing tests to python-bioimage-io
necessary changes to make tests run :)

* changed imports from tiktorch (of course)
* removed `REMOVE_BATCH_DIM`, `ADD_BATCH_DIM`
  those are currently commented out - why?

+++ bioimageio/core/prediction_pipeline/_preprocessing.py
@@ -34,9 +34,7 @@ def zero_mean_unit_variance(tensor: xr.DataArray, axes=None, eps=1.0e-6, mode="p
 
     ret = (tensor - mean) / (std + 1.0e-6)
 
-    # monkey patch: maks sure we don't change dtype
-    # todo: allow preprocessing to change dtype?
-    return ret.astype(tensor.dtype)
+    return ret
 
 
 def binarize(tensor: xr.DataArray, *, threshold) -> xr.DataArray:
